+++ src/loadsave.h
@@ -32,8 +32,6 @@ extern MAZELEVEL *mapmem;
 extern CSPRITE *csprite;
 
 const char* RAP_InitLoadSave(void);
-const char* RAP_SetupFilename(void);
-void RAP_WriteDefaultSetup(void);
 const char* RAP_GetPath(void);
 int RAP_CheckFileInPath(const char* filename);